.
Former-commit-id: 9ccecdc1673a4ffb2e845039028ce58ca057c854

+++ core/src/diagnostics/recover_components/recov_hor_par_dual.c
@@ -1,7 +1,7 @@
 #include "../../enum_and_typedefs.h"
 #include <stdio.h>
 
-int recov_hor_par_dual(Dual_vector_field in_field, long layer_index, long h_index, double *component, Grid *grid)
+int recov_hor_par_dual(Dual_vector_field in_field, int layer_index, int h_index, double *component, Grid *grid)
 {
     *component = 0;
     for (int i = 0; i < 2; ++i)
