+++ core/src/diagnostics/recover_components/recov_hor_par_pri.c
@@ -1,6 +1,6 @@
 #include "../../enum_and_typedefs.h"
 
-int recov_hor_par_pri(Vector_field in_field, long layer_index, long h_index, double *component, Grid *grid)
+int recov_hor_par_pri(Vector_field in_field, int layer_index, int h_index, double *component, Grid *grid)
 {
     *component = 0;
     for (int i = 0; i < 10; ++i)
